+++ okhttp/src/main/java/okhttp3/internal/tls/InsecureTrustManager.kt
@@ -1,0 +1,27 @@
+/*
+ * Copyright (C) 2020 Square, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package okhttp3.internal.tls
+
+import java.security.cert.X509Certificate
+import javax.net.ssl.X509TrustManager
+
+internal object InsecureTrustManager : X509TrustManager {
+  override fun checkClientTrusted(chain: Array<X509Certificate>, authType: String) {}
+
+  override fun checkServerTrusted(chain: Array<X509Certificate>, authType: String) {}
+
+  override fun getAcceptedIssuers(): Array<X509Certificate> = arrayOf()
+}
