+++ okhttp/src/main/java/okhttp3/internal/tls/TrustManagerBridge.kt
@@ -13,32 +13,20 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package okhttp3.tls
+package okhttp3.internal.tls
 
 import android.annotation.SuppressLint
-import java.security.cert.X509Certificate
 import javax.net.ssl.X509ExtendedTrustManager
 import javax.net.ssl.X509TrustManager
 import okhttp3.internal.platform.Platform
-import okhttp3.tls.internal.OkHttpTrustManagerAndroid
-import okhttp3.tls.internal.OkHttpTrustManagerJvm
-import okhttp3.tls.internal.TrustManagerOverride
 import org.codehaus.mojo.animal_sniffer.IgnoreJRERequirement
 
-object InsecureTrustManager : X509TrustManager {
-  override fun checkClientTrusted(chain: Array<X509Certificate>, authType: String) {}
-
-  override fun checkServerTrusted(chain: Array<X509Certificate>, authType: String) {}
-
-  override fun getAcceptedIssuers(): Array<X509Certificate> = arrayOf()
-}
-
-object OkHttpTrustManager {
-  class Builder(var delegate: X509TrustManager) {
+object TrustManagerBridge {
+  class Builder(var default: X509TrustManager? = null) {
     private var overrides: MutableList<TrustManagerOverride> = mutableListOf()
 
-    fun delegate(delegate: X509TrustManager) = apply {
-      this.delegate = delegate
+    fun default(default: X509TrustManager) = apply {
+      this.default = default
     }
 
     fun hostOverride(hostName: String, trustManager: X509TrustManager) = apply {
@@ -51,18 +39,15 @@ object OkHttpTrustManager {
           InsecureTrustManager))
     }
 
-    fun hostCertificates(hostName: String, handshakeCertificates: HandshakeCertificates) = apply {
-      TODO("what should the API look like for jumping from HandshakeCertificates")
-    }
-
     @IgnoreJRERequirement
     @SuppressLint("NewApi")
     fun build(): X509TrustManager {
+      val defaultTrustManager = default ?: Platform.get().platformTrustManager()
+
       return if (Platform.get().isAndroid) {
-        OkHttpTrustManagerAndroid(delegate, overrides.toList())
+        OkHttpTrustManagerAndroid(defaultTrustManager, overrides.toList())
       } else {
-        OkHttpTrustManagerJvm(
-            delegate as X509ExtendedTrustManager, overrides.toList())
+        OkHttpTrustManagerJvm(defaultTrustManager as X509ExtendedTrustManager, overrides.toList())
       }
     }
   }
